working on read/write cgi
