--- cgi-eval-text.txt
@@ -1,8 +1,0 @@
-
-- The server is working fine using a CGI.
-- The CGI should be run in the correct directory for relative path file access.
-- With the help of the students you should check that everything is working properly. You have to test the CGI with the "GET" and "POST" methods.
-- You need to test with files containing errors to see if the error handling works properly. You can use a script containing an infinite loop or an error; you are free to do whatever tests you want within the limits of acceptability that remain at your discretion. The group being evaluated should help you with this.
-
-The server should never crash and an error should be visible in case of a problem.
-
