--- todo
@@ -1,78 +1,0 @@
-// request parser
-1. Data lezen in stukken in een while(1) loop 
-2. parse de header en check voor \r\n\r\n. Met dit kan je ervan uitgaan dat de request header compleet is
-3. Kijk naar de content-length. als het er is, is er een body. De length geeft aan hoeveel data er in de body zit en is een indicatie hoeveel data je al hebt ontvangen
-4. Parse de request body (alleen bij post requests). Blijf de data in chuncks lezen, wanneer de binnengekomende length matched met de content-length, is de gehele request compleet.
-
-// cgi and post
-pipe aanmaken en de write kant toevoegen aan epoll
-wanneer de epoll klaar is kan het een request handelen
-1. CGi request binnen
-2. de body duppen naar STDIN
-3. excev naar een python script dat de inpout naar STDOUT convert
-4. van STDOUT naar response naar de client
-
-// juiste response sturen
-
-check content length for post keep looping till no bytes read
-
-alles onder boundary copy naar file naam filename in root
-
-goed!
-
-Kijk of hij loopt tot nieuw line;
-if so parse alles erna naar _body 
-kijk ff hoe dit werkt met string allocation en max sizes blabla
-daarna pars naar file
-
-///////////////////////////////////////////////////////////////////////
-
-for (int i = 0; i < numEvents; ++i) {
-    if (events[i].data.fd == serverSocket) {
-        // Handle new connections (already handled in the previous example)
-    } else {
-        // Handle data from clients
-        if (events[i].events & EPOLLIN) {
-            // Read data from the client
-            char buffer[1024];
-            int bytesRead = recv(events[i].data.fd, buffer, sizeof(buffer), 0);
-
-            if (bytesRead > 0) {
-                // Process the received data (you can replace this with your own logic)
-                buffer[bytesRead] = '\0';
-                std::cout << "Received data from client " << events[i].data.fd << ": " << buffer << std::endl;
-            } else if (bytesRead == 0) {
-                // Connection closed by the client
-                std::cout << "Client " << events[i].data.fd << " disconnected." << std::endl;
-                close(events[i].data.fd);
-                epoll_ctl(epollFd, EPOLL_CTL_DEL, events[i].data.fd, nullptr);
-            } else {
-                // Error in recv
-                perror("Error in recv");
-                close(events[i].data.fd);
-                epoll_ctl(epollFd, EPOLL_CTL_DEL, events[i].data.fd, nullptr);
-            }
-        }
-
-        if (events[i].events & EPOLLOUT) {
-            // Write data to the client
-            const char* message = "Hello, client!";
-            int bytesWritten = send(events[i].data.fd, message, strlen(message), 0);
-
-            if (bytesWritten == -1) {
-                // Error in send
-                perror("Error in send");
-                close(events[i].data.fd);
-                epoll_ctl(epollFd, EPOLL_CTL_DEL, events[i].data.fd, nullptr);
-            } else {
-                std::cout << "Sent data to client " << events[i].data.fd << ": " << message << std::endl;
-            }
-        }
-    }
-}
-/////////////////////////////
-
-FIX METHODS
-empty when in is path and method allowed
-after posting
-autoindex directory listing
